pyproject.toml
[tool.poetry]
name = "uzumaki-slack-chatgpt"
version = "0.1.0"
description = ""
authors = ["cfiken <cfiken@gmail.com>"]
readme = "README.md"
packages = [{include = "slack_app"}]

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.16.1"
python-dotenv = "^0.21.0"
openai = "^0.27.2"
langchain = "^0.0.123"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
